Rename sugar module -> resource_block
