+++ polar-core/src/lib.rs
@@ -23,12 +23,12 @@ mod numerics;
 pub mod parser;
 mod partial;
 pub mod polar;
+mod resource_block;
 mod rewrites;
 pub mod roles_validation;
 pub mod rules;
 mod runnable;
 mod sources;
-mod sugar;
 pub mod terms;
 pub mod traces;
 mod visitor;
